testgame/game1/requirements.txt
pygame==2.4.0
streamlit==1.19.0
Pillow==8.4.0
